@@ -21,35 +21,32 @@
   },
   "dependencies": {
     "@codegouvfr/react-dsfr": "^0.21.0",
-    "@emotion/react": "^11.9.0",
-    "@emotion/styled": "^11.8.1",
     "@gouvfr/dsfr": "1.8.5",
     "@mui/icons-material": "^5.5.1",
     "@mui/material": "^5.5.2",
     "@reduxjs/toolkit": "^1.8.1",
-    "@types/react-helmet": "^6.1.6",
     "axios": "^0.26.1",
     "date-fns": "^2.28.0",
     "eslint": "^8.11.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-jest": "^26.1.2",
-    "formik": "2.2.9",
+    "formik": "^2.2.9",
     "html-templates": "workspace:*",
     "http-client": "workspace:*",
     "ramda": "^0.28.0",
-    "react": "18",
+    "react": "^18.2.0",
     "react-design-system": "workspace:*",
-    "react-dom": "18",
-    "react-helmet": "^6.1.0",
+    "react-dom": "^18.2.0",
+    "react-helmet-async": "^1.3.0",
     "react-redux": "^7.2.8",
     "redux": "^4.2.0",
     "redux-observable": "^2.0.0",
     "rxjs": "^7.5.5",
     "shared": "workspace:*",
     "tss-react": "^4.5.2",
-    "type-route": "^0.6.0",
+    "type-route": "^1.0.1",
     "uuid": "^8.3.2",
-    "zod": "^3.14.2"
+    "zod": "^3.20.2"
   },
   "devDependencies": {
     "@babel/core": "^7.17.9",
@@ -58,18 +55,18 @@
     "@types/jest": "^27.4.1",
     "@types/node": "^17.0.31",
     "@types/ramda": "^0.28.1",
-    "@types/react": "^17.0.41",
-    "@types/react-dom": "^17.0.14",
-    "@types/react-redux": "^7.1.23",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
+    "@types/react-redux": "^7.1.25",
     "@types/uuid": "^8.3.4",
-    "@vitejs/plugin-react": "^2.1.0",
+    "@vitejs/plugin-react": "^2.2.0",
     "autoprefixer": "^10.4.4",
     "babel-jest": "^27.5.1",
     "jest": "^27.5.1",
     "postcss": "^8.4.12",
     "sass": "^1.55.0",
     "tailwindcss": "^3.0.23",
-    "typescript": "^4.6.2",
+    "typescript": "^4.8.2",
     "vite": "^3.1.4"
   }
 }
